feature: Register a sensor node reading in the database / Update: uuid validation
- feature: Register a sensor node reading in the database
Register a sensor node reading in the database. Checks if the sensor node informed
exists. Check if reading was already registered in the database. Check if collected
date is valid (not a date in the future). If reading meets the conditions, it is stored
in the database.

-----------------------------------------------------------------------------------
Update: UUID validation

When the method isUUID receive undefined or null, it returns a TypeError instead of
a false result (invalid uuid). So the test was modified to only runs when there
is a value passed to uuid, otherwise returns false.
Cheking the validator.js library in github to verify the implementation of isUUID,
it looks like it is working as intended since the interface expects to receive a
string as UUID, otherwise it throws the TypeError exception.